requirements.txt
requests
pandas
numpy
matplotlib
scikit-learn
streamlit
plotly
statsmodels
xgboost
python-dotenv
alpaca-py